+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.26.0] - 2020-04-25
+
+### Added
+
+- Added Electra model support for sequence classification (binary, multiclass, multilabel)
+- Added Electra model support for question answering
+- Added Roberta model support for question answering
+
+### Changed
+
+- Reduced logger messages during question answering evaluation
+
 ## [0.25.0] - 2020-04-24
 
 ### Added
@@ -609,7 +621,9 @@ Model checkpoint is now saved for all epochs again.
 
 - This CHANGELOG file to hopefully serve as an evolving example of a standardized open source project CHANGELOG.
 
-[0.25.0]: https://github.com/ThilinaRajapakse/simpletransformers/compare/445d386...HEAD
+[0.25.6]: https://github.com/ThilinaRajapakse/simpletransformers/compare/aa8e6a6...HEAD
+
+[0.25.0]: https://github.com/ThilinaRajapakse/simpletransformers/compare/445d386...aa8e6a6
 
 [0.24.9]: https://github.com/ThilinaRajapakse/simpletransformers/compare/52fea69...445d386